Update test.py
This code will help you understand how to set up a function, a variable, an if/if-else statement, and an input. These are four basic snippets of code that are used in app development, website creation, and many more code implementations. Practice this code section and try to master it. Have fun, study, hard, and keep up the inspired work.